Address Review Comments

+++ core/unit_test/TestMDSpanConversion.hpp
@@ -513,46 +513,46 @@ TEST(TEST_CATEGORY, view_mdspan_conversion_with_stride) {
                                    Kokkos::pair{2, 38});
     auto sub_mds = sub_v.to_mdspan();
     Kokkos::View<int ***, Kokkos::LayoutLeft> sub_v2(sub_mds);
-    ASSERT_EQ(sub_v.extent(0), 10);
-    ASSERT_EQ(sub_v.extent(1), 40);
-    ASSERT_EQ(sub_v.extent(2), 36);
-    ASSERT_EQ(sub_v.stride(0), 1);
-    ASSERT_EQ(sub_v.stride(1), 20);
-    ASSERT_EQ(sub_v.stride(2), 800);
-    ASSERT_EQ(sub_mds.extent(0), 10);
-    ASSERT_EQ(sub_mds.extent(1), 40);
-    ASSERT_EQ(sub_mds.extent(2), 36);
-    ASSERT_EQ(sub_mds.stride(0), 1);
-    ASSERT_EQ(sub_mds.stride(1), 20);
-    ASSERT_EQ(sub_mds.stride(2), 800);
-    ASSERT_EQ(sub_v2.extent(0), 10);
-    ASSERT_EQ(sub_v2.extent(1), 40);
-    ASSERT_EQ(sub_v2.extent(2), 36);
-    ASSERT_EQ(sub_v2.stride(0), 1);
-    ASSERT_EQ(sub_v2.stride(1), 20);
-    ASSERT_EQ(sub_v2.stride(2), 800);
+    ASSERT_EQ(static_cast<int>(sub_v.extent(0)), 10);
+    ASSERT_EQ(static_cast<int>(sub_v.extent(1)), 40);
+    ASSERT_EQ(static_cast<int>(sub_v.extent(2)), 36);
+    ASSERT_EQ(static_cast<int>(sub_v.stride(0)), 1);
+    ASSERT_EQ(static_cast<int>(sub_v.stride(1)), 20);
+    ASSERT_EQ(static_cast<int>(sub_v.stride(2)), 800);
+    ASSERT_EQ(static_cast<int>(sub_mds.extent(0)), 10);
+    ASSERT_EQ(static_cast<int>(sub_mds.extent(1)), 40);
+    ASSERT_EQ(static_cast<int>(sub_mds.extent(2)), 36);
+    ASSERT_EQ(static_cast<int>(sub_mds.stride(0)), 1);
+    ASSERT_EQ(static_cast<int>(sub_mds.stride(1)), 20);
+    ASSERT_EQ(static_cast<int>(sub_mds.stride(2)), 800);
+    ASSERT_EQ(static_cast<int>(sub_v2.extent(0)), 10);
+    ASSERT_EQ(static_cast<int>(sub_v2.extent(1)), 40);
+    ASSERT_EQ(static_cast<int>(sub_v2.extent(2)), 36);
+    ASSERT_EQ(static_cast<int>(sub_v2.stride(0)), 1);
+    ASSERT_EQ(static_cast<int>(sub_v2.stride(1)), 20);
+    ASSERT_EQ(static_cast<int>(sub_v2.stride(2)), 800);
   }
   {
     // layout_right_padded<dynamic_extent> has a custom stride for
     // stride(rank-2) LayoutRight has a custom stride for stride(0) That means
-    // the "padding" only mateches up for Rank-2 Views
+    // the "padding" only matches up for Rank-2 Views
     Kokkos::View<int **, Kokkos::LayoutRight> source("S", 20, 40);
     auto sub_v =
         Kokkos::subview(source, Kokkos::pair{5, 15}, Kokkos::pair{2, 38});
     auto sub_mds = sub_v.to_mdspan();
     Kokkos::View<int **, Kokkos::LayoutRight> sub_v2(sub_mds);
-    ASSERT_EQ(sub_v.extent(0), 10);
-    ASSERT_EQ(sub_v.extent(1), 36);
-    ASSERT_EQ(sub_v.stride(0), 40);
-    ASSERT_EQ(sub_v.stride(1), 1);
-    ASSERT_EQ(sub_mds.extent(0), 10);
-    ASSERT_EQ(sub_mds.extent(1), 36);
-    ASSERT_EQ(sub_mds.stride(0), 40);
-    ASSERT_EQ(sub_mds.stride(1), 1);
-    ASSERT_EQ(sub_v2.extent(0), 10);
-    ASSERT_EQ(sub_v2.extent(1), 36);
-    ASSERT_EQ(sub_v2.stride(0), 40);
-    ASSERT_EQ(sub_v2.stride(1), 1);
+    ASSERT_EQ(static_cast<int>(sub_v.extent(0)), 10);
+    ASSERT_EQ(static_cast<int>(sub_v.extent(1)), 36);
+    ASSERT_EQ(static_cast<int>(sub_v.stride(0)), 40);
+    ASSERT_EQ(static_cast<int>(sub_v.stride(1)), 1);
+    ASSERT_EQ(static_cast<int>(sub_mds.extent(0)), 10);
+    ASSERT_EQ(static_cast<int>(sub_mds.extent(1)), 36);
+    ASSERT_EQ(static_cast<int>(sub_mds.stride(0)), 40);
+    ASSERT_EQ(static_cast<int>(sub_mds.stride(1)), 1);
+    ASSERT_EQ(static_cast<int>(sub_v2.extent(0)), 10);
+    ASSERT_EQ(static_cast<int>(sub_v2.extent(1)), 36);
+    ASSERT_EQ(static_cast<int>(sub_v2.stride(0)), 40);
+    ASSERT_EQ(static_cast<int>(sub_v2.stride(1)), 1);
   }
 }
 }  // namespace
